added controller for kitchen
